+++ packages/twake-mui/src/lib/overrides/dialogOverrides.ts
@@ -8,7 +8,8 @@ export const dialogOverrides = (): Components['MuiDialog'] => {
       paper: {
         borderRadius: radius.md,
         boxShadow:
-          '0 1px 3px 0 rgba(0, 0, 0, 0.30), 0 4px 8px 3px rgba(0, 0, 0, 0.15)'
+          '0 1px 3px 0 rgba(0, 0, 0, 0.30), 0 4px 8px 3px rgba(0, 0, 0, 0.15)',
+        maxWidth: '570px'
       }
     }
   }